Change double to float to descrease the amount of memory in used.

+++ My_first_Cpp_game/game.h
@@ -8,4 +8,4 @@
 #define isReleased(b) (!input->buttons[b].isDown && input->buttons[b].changed)
 
 /* FUNCTION */
-void stimulateGame(Input* input, double deltaTime);
+void stimulateGame(Input* input, float deltaTime);
